Add script to copy compose images to our registry
When building a compose App, we need a copy of the original images. They
may be hosted in a private registry, or a temporary location. To ensure
they are always hosted with us, clone them upon building a compose App.

After cloning, replace the reference in the original App manifest to
reference our official repository.